requirements.txt
dictalchemy==0.1.2.7
flask==1.0.2
Flask-Script==2.0.6
Flask-Migrate==2.3.1
mysqlclient==1.4.1
redis==2.10.6
requests==2.21.0
SQLAlchemy==1.3.0
unittest2==1.1.0
google-api-python-client==1.7.7
python-dateutil==2.7.5
